\name{rgpg-package}
\alias{rgpg-package}
\alias{rgpg}
\docType{package}
\title{
\packageTitle{rgpg}
}
\description{
\packageDescription{rgpg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rgpg}
\packageIndices{rgpg}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rgpg}

Maintainer: \packageMaintainer{rgpg}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
